API + ENH : move ctf + psf functions + optim
