+++ examples/inverse/plot_mne_crosstalk_function.py
@@ -1,12 +1,10 @@
 """
-===========================
-Compute cross-talk functions (CTFs) for labels
-for different linear inverse operators
-===========================
+===================================================================
+Compute cross-talk functions (CTFs) for labels for MNE/dSPM/sLORETA
+===================================================================
 
 CTFs are computed for four labels in the MNE sample data set
-for three linear inverse operators (MNE, dSPM, sLORETA).
-CTFs are saved as STC files for visualization in mne_analyze.
+for linear inverse operators (MNE, dSPM, sLORETA).
 CTFs describe the sensitivity of a linear estimator (e.g. for
 one label) to sources across the cortical surface. Sensitivity
 to sources outside the label is undesirable, and referred to as
@@ -23,18 +21,15 @@ import mne
 from mne.datasets import sample
 from mne.minimum_norm import cross_talk_function, read_inverse_operator
 
-
-## Example how to compute CTFs
-
 data_path = sample.data_path()
+subjects_dir = data_path + '/subjects/'
 fname_fwd = data_path + '/MEG/sample/sample_audvis-meg-eeg-oct-6-fwd.fif'
 fname_inv = data_path + '/MEG/sample/sample_audvis-meg-oct-6-meg-inv.fif'
 fname_evoked = data_path + '/MEG/sample/sample_audvis-ave.fif'
-
-fname_label = [data_path + '/MEG/sample/labels/Vis-rh.label',
-               data_path + '/MEG/sample/labels/Vis-lh.label',
-               data_path + '/MEG/sample/labels/Aud-rh.label',
-               data_path + '/MEG/sample/labels/Aud-lh.label']
+fname_label = [data_path + '/MEG/sample/labels/Aud-rh.label',
+               data_path + '/MEG/sample/labels/Aud-lh.label',
+               data_path + '/MEG/sample/labels/Vis-rh.label',
+               data_path + '/MEG/sample/labels/Vis-lh.label']
 
 # In order to get leadfield with fixed source orientation,
 # read forward solution with fixed orientations
@@ -45,13 +40,26 @@ labels = [mne.read_label(ss) for ss in fname_label]
 
 inverse_operator = read_inverse_operator(fname_inv)
 
-fname_stem = 'ctf'
-for method in ('MNE', 'dSPM', 'sLORETA'):
-    stc_ctf, label_singvals = cross_talk_function(inverse_operator, forward,
-                                                  labels, method=method,
-                                                  lambda2=1 / 9., signed=True,
-                                                  mode='svd', n_svd_comp=3)
+# regularisation parameter
+snr = 3.0
+lambda2 = 1.0 / snr ** 2
+method = 'dSPM'  # can be 'MNE' or 'sLORETA'
 
-    fname_out = fname_stem + '_' + method
-    print("Writing CTFs to files %s" % fname_out)
-    stc_ctf.save(fname_out)
+stc_ctf, _ = cross_talk_function(inverse_operator, forward, labels,
+                                 method=method, lambda2=lambda2, signed=True,
+                                 mode='svd', n_svd_comp=3)
+
+fmax = stc_ctf.data[:, 0].max()
+fmid = fmax / 2.
+fmin = 0.
+
+time_label = "Label %d"
+
+brain = stc_ctf.plot(subject='sample', surface='inflated', hemi='rh',
+                     subjects_dir=subjects_dir, time_label=time_label,
+                     fmin=fmin, fmid=fmid, fmax=fmax)
+
+# Save CTFs for visualization in mne_analyze.
+# fname_out = fname_stem + '_' + method
+# print("Writing CTFs to files %s" % fname_out)
+# stc_ctf.save(fname_out)
