+++ mne/minimum_norm/__init__.py
@@ -3,6 +3,7 @@
 from .inverse import (read_inverse_operator, apply_inverse,
                       apply_inverse_raw, make_inverse_operator,
                       apply_inverse_epochs, write_inverse_operator,
-                      compute_rank_inverse, point_spread_function, cross_talk_function)
+                      compute_rank_inverse)
+from psf_ctf import point_spread_function, cross_talk_function
 from .time_frequency import (source_band_induced_power, source_induced_power,
                              compute_source_psd, compute_source_psd_epochs)
